Version 1.6.28 (#17)
* Fix perf for dng

* Fix nikon Preview

* Add doc

* Fix Nikon preview

* Fix

* Fix raw decoder

* New version

* Add a donate button

* Small imporovement for low memory device

* Change app name

* Clean up

* Fix a null exception

* Quick fix fo demosaic

* First fix for RW2

* Add support for RAW2 thumbnail

* Clean up

* Clean up for RAW v1

* Fix for Nikon d2x (white balance still not correct)

* Update nuget packages

+++ Source/RawParser/Model/Effect/Demosaic/SSDDemosaic.cs
@@ -1,9 +1,5 @@
 ﻿using RawNet;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RawEditor.Effect
 {
@@ -87,7 +83,7 @@ namespace RawEditor.Effect
         static double LUT(double dif, double[] lut)
         {
             if (dif >= (float)LUTMAXM1) return 0.0;
-            int x = (int)Math.Floor((double)dif * (float)LUTPRECISION);
+            int x = (int)Math.Floor(dif * (float)LUTPRECISION);
             double y1 = lut[x];
             double y2 = lut[x + 1];
             return y1 + (y2 - y1) * (dif * LUTPRECISION - x);
